+++ src/main/java/uk/gov/ons/census/fwmt/jobservice/Application.java
@@ -26,23 +26,4 @@ public class Application {
     SpringApplication.run(Application.class, args);
   }
 
-  @Value("#{'${logging.profile}' == 'CLOUD'}")
-  private boolean useJsonLogging;
-
-  @PostConstruct
-  public void initJsonLogging() {
-    HashMap<Class<?>, Function<Object, String>> customMappers = new HashMap<>();
-    customMappers.put(LocalTime.class, Object::toString);
-    customMappers.put(LocalDateTime.class, Object::toString);
-
-    LoggingConfigs configs;
-
-    if (useJsonLogging) {
-      configs = LoggingConfigs.builder().customMapper(customMappers).build().useJson();
-    }
-    else {
-      configs = LoggingConfigs.builder().customMapper(customMappers).build();
-    }
-    LoggingConfigs.setCurrent(configs);
-  }
 }
